backend/requirements.txt
fastapi
uvicorn
twilio
python-dotenv
pydantic